requirements.txt
alembic==1.17.2
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
dotenv==0.9.9
Flask==3.1.2
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
openai==2.8.1
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
psycopg2==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
rpds-py==0.29.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
